agent_system_b/requirements.txt
# Agent System B — Google ADK Independent Valuation
google-adk
litellm>=1.75.5
pydantic
python-dotenv
qdrant-client
openai
pandas
httpx
fastapi
uvicorn
